todas as peças já na nova 'arquitetura'
