+++ tabuleiro.py
@@ -1,7 +1,5 @@
 import pygame 
-from pygame.locals import *
-from classes.pawn import Pawn
-from importador import *
+
 
 class Tabuleiro:
     def __init__(self,tela):
@@ -21,38 +19,38 @@ class Tabuleiro:
                     self.__desenhar_linha(linha, coluna, 'dark green', 'white')
 
 
-    def desenhar_peças_pretas(self):
-        black_pieces_group=pygame.sprite.Group()
-        for peça_preta in black_pieces:
-            black_pieces_group.add(peça_preta)
+    # def desenhar_peças_pretas(self):
+    #     black_pieces_group=pygame.sprite.Group()
+    #     for peça_preta in black_pieces:
+    #         black_pieces_group.add(peça_preta)
 
-        black_pieces_group.draw(self.tela)
-        black_pieces_group.update()
+    #     black_pieces_group.draw(self.tela)
+    #     black_pieces_group.update()
 
-        #desenhando peão
+    #     #desenhando peão
 
 
-    def desenhar_peças_brancas(self):
-        white_pieces_group=pygame.sprite.Group()
-        for peça_branca in white_pieces:
-            white_pieces_group.add(peça_branca)
+    # def desenhar_peças_brancas(self):
+    #     white_pieces_group=pygame.sprite.Group()
+    #     for peça_branca in white_pieces:
+    #         white_pieces_group.add(peça_branca)
 
-        white_pieces_group.draw(self.tela)
-        white_pieces_group.update()
+    #     white_pieces_group.draw(self.tela)
+    #     white_pieces_group.update()
         
 
-        #desenhando peão
-        x=0
-        y=self.tam_tabuleiro
-        y= y - self.tamanho_quadrado
-        white_pawns_group=pygame.sprite.Group()
-        for i in range(8):
-            white_pawns.append(Pawn((x,y),"white"))
-            white_pawns_group.add(white_pawns[-1])
-            x= x+ self.tamanho_quadrado
+    #     #desenhando peão
+    #     x=0
+    #     y=self.tam_tabuleiro
+    #     y= y - self.tamanho_quadrado
+    #     white_pawns_group=pygame.sprite.Group()
+    #     for i in range(8):
+    #         white_pawns.append(Pawn((x,y),"white"))
+    #         white_pawns_group.add(white_pawns[-1])
+    #         x= x+ self.tamanho_quadrado
 
-        white_pawns_group.draw(self.tela)
-        white_pawns_group.update()
+    #     white_pawns_group.draw(self.tela)
+    #     white_pawns_group.update()
 
 
     #metodos privados
